refactor: move DeleteButton tu ui
