+++ src/features/vocabulary/components/Meaning/Meaning.tsx
@@ -1,9 +1,8 @@
 import React, { FC, useState } from 'react'
 import { Button } from 'antd'
 import { PlusCircleOutlined } from '@ant-design/icons'
-import { DeleteButton } from '../DeleteButton'
 import { Translation } from '../Translation'
-import { List } from '../../../../ui'
+import { List, DeleteButton } from '../../../../ui'
 import { AddTranslationForm } from '../AddTranslationForm'
 import { ITranslation } from '../../../../models/ITranslation'
 import { IMeaning } from '../../../../models/IMeaning'
