+++ src/ui/index.ts
@@ -2,3 +2,4 @@ export { List } from './List'
 export { Input } from './Input'
 export { SubmitButton } from './SubmitButton'
 export { Form } from './Form'
+export { DeleteButton } from './DeleteButton'
